v5.0.0 - Drop analysis tools from Reffy
This major update drops the analysis bits that have now been moved to Strudy,
as envisioned in #747.

On top of the analysis code per se, this update also drops the `crawl-and-study`
tool that applied Reffy and Strudy in combination to create Webref reports.
The same thing can be achieved through a short set of commands, which typically
appear in the update jobs run in w3c/webref and w3c/webref-analysis.

The update drops the `npm run` commands that were referencing `crawl-and-study`.

The `check-specs` utility was moved to a dedicated repo. The future of this
utility is uncertain. It had initially been created to run a "Reffy service"
that would allow editors to check a local version of their spec on-the-fly. It
has never been used in practice. The "backup" repo is at:
https://github.com/tidoust/check-specs

The `check-missing-dfns` utility should go too since it now belongs to Strudy.
Problem is `generate-idlnames` needs a couple of its functions, so code remains
in Reffy for now. To be addressed in a future update!

The update also exposes functions that Strudy and `check-specs` depend on more
directly to avoid clumsy `require` statements à la
`require('reffy/src/lib/util')`.