Serialize Puppeteer setup/teardowns
With the current architecture, there can be only one instance of Puppeteer
running at the same time, and since `crawlList` and `crawlSpecs` are the ones
which call `setupBrowser` and `teardownBrowser`, callers must not be able to
run multiple crawlers in parallel.

Problem is that it is possible to call `crawlList` or `crawlSpecs` in parallel.
This does not happen "in real life" when Reffy runs, but that typically happens
in tests, as they call the crawl functions multiple times. This meant that
browser instances were created but not closed, which in turn might explain why
tests time out from time to time.

This update makes sure that crawler runs are serialized, in other words that
there can be only one Puppeteer instance at any given time.