Add parsed IDL generator tool
The tool contains the logic to generate parsed IDL structures from the raw IDL.
It takes a crawl report as input and creates JSON files in an `idlparsed`
folder.

The crawler now uses the functions exported by the tool to generate the
`idlparsed` extracts.

The tool is needed to creates curated versions of reports and in particular to
re-generate the `idlparsed` extracts once IDL patches have been applied to the
crawled IDL.

This update includes some tests on the generation.