Separate raw and parsed IDL in in-memory spec structure
The in-memory representation of a crawled spec now properly separates the raw
IDL, stored in an `idl` property, and the parsed IDL, stored in an `idlparsed`
property.

This is a breaking change for projects that use `expandCrawlResults` to load a
crawl report into memory, and a major bump will be needed.

See discussion in https://github.com/w3c/reffy/pull/814#discussion_r779451242