Release new major version of Reffy - v6.0.0
New major version is needed due to the recent commit that separates the raw IDL
and the parsed IDL structure in the in-memory model of a crawled spec:
https://github.com/w3c/reffy/commit/747df3cd8b1777f5e5398069998f4fbde534e39c

This change does not modify generated extracts but means that the spec object
returned by the `expandCrawlResults` function is slightly different. The parsed
IDL structure used to be found under `spec.idl` and the raw IDL in
`spec.idl.idl`. The parsed IDL structure now appears under `spec.idlparsed` and
the raw IDL in `spec.idl`.

Non-breaking changes on top of usual dependency version bumps:
- New tool: parsed IDL generator (in `src/cli`)
- Bug fix: Serialize Puppeteer setup/teardowns
- Improvement: Improve network interception test logic