Expose idlnames and idlparsed utility functions in package
This makes it easier to reuse these functions from within other packages.

The update also adjusts the `expandCrawlResults` function to convert CRLF to LF
when it loads text files, as Git may auto-convert LF to CRLF on checkout on
Windows machines (that does not create any error per se, but we'll want to reuse
the expanded IDL to create multiline IDL fragments in parsed IDL constructs, and
it seems good to make sure that all platforms generate the exact same IDL
fragments).