Improve fallback logic
This update moves the fallback logic to the main crawl function as the CLI
reports results to the console in the absence of an `--output` parameter.

The match is now also made on the initial URL rather than on the shortname
(crawl tests may not always produce a shortname), the initial URL is guaranteed
to exist.

The crawler reports the `error` property even when it uses fallback data, so
that someone reading the crawl result may understand that fallback data was used
and why it was used.

A new test was added to check the fallback mechanism.

Also note the code was adjusted to stop modifying the options object sent to
functions, and to rather make the updates on a shallow copy of the object.