Fix fallback logic to actually handle extracts
Replacing data with fallback data is good provided that we replace data with
similar data. The save function expects expanded results and cannot process a
spec result that already contains links to extracts without the expanded
content.

In any case, the output of the crawl function needs to be the same regardless of
whether the fallback option was set. Also, the fallback folder may not be the
same as the output folder, in which case relative links taken from the fallback
crawl essentially are invalid.

This update expands the fallback data when it gets used. The save function can
then process the spec result as usual, and create the right links.

Tests completed with a test on extracts.