Link to series extracts from one spec, make idlparsed per series (#851)
This fixes #844. The crawler kept a link to series extracts (CSS and IDL) from
all specs in the series. This is clumsy at best, and basically wrong as the
extracts may not represent the actual content defined in that spec. The crawler
no longer does that.

The `idlparsed` folder is built from the `idl` folder but was per spec instead
of per series. For consistency (and to avoid creating duplicate specs under
`idlparsed`), the `idlparsed` folder now contains extracts per spec series.