IDL extraction: drop check on NoInterfaceObject
Fix for #87. The extraction code took for granted that an interface defined with
`NoInterfaceObject` could not be referenced, which is wrong.

In practice, the `NoInterfaceObject` extended attribute was replaced by
`LegacyNoInterfaceObject` some time ago, meaning that the condition always
passed in practice, so bug was already fixed.

FWIW, the only interfaces that still use `LegacyNoInterfaceObject` are WebGL
extensions.