Well-known Web IDL types: Drop Error, RegExp and DOMException (#931)
PR #771 took care of most Web IDL types that are defined in the Web IDL spec as
actual IDL interfaces, except `DOMException`. It appears 24 times across all
specs, that seems like a good dependency to track as well (and not too noisy).

`Error` and `RegExp` are not valid Web IDL types, so not sure why they were
added to the list of well-known Web IDL types in the first place!

This would fix #516. Automation seems overkill for that.