Add module to extract info on event types and their associated event interfaces (#952)
This adds a new default crawling module to Reffy that parses specs and extracts information
on event types, their associated event interfaces, the interfaces on which they can fire when
known, as well as on whether they bubble. Extracts also contain some information about
where the crawler found the event info in the spec.

Events extracts are created in an `events` folder.