Re-write CSS extraction module based on "data-dfn-type" (#976)
This re-writes the CSS extraction module to rely on the `data-dfn-type`
attribute. Only production definitions marked with `type` or `function` get
extracted.

As an exception to the rule, `<pre>` tags are parsed separately because,
while CSS specs consistently use dfns in other forms of production rules, they
sometimes mix dfns and references in `<pre>` tags, as in:
https://drafts.csswg.org/css-values-4/#calc-syntax

When definition appears in prose, the paragraph is copied.