Fix CSS production rules split logic
Fixes #983 by making the production rules logic more resilient to the
possibility that multiple production rules get defined in prose at the same
level or with one being a child of the other.

The fix also makes the regular expression used to split multiple production
rules more robust to cases where the equal sign appears on a different line
from the type definition.