Add 1mn timeout for Respec generation
Respec will never be able to generate a spec that has a syntax error in its
Respec configuration. In such cases, Respec will never resolve the "ready"
promise.

This update adds a 1mn timeout to make sure that the crawler does not get
stuck waiting forever.

(Updated prompted by a config bug in the PNG spec that makes the crawler
loop forever)