Fix exported functions crawlSpecs and getInterfaceTreeInfo
The `getTreeInfo` entry point should have been renamed into
`getInterfaceTreeInfo`.

The `crawlSpecs` function targeted the `crawlList` crawl function instead of
`crawlSpecs`. I'm not clear that was intended and it would be easier for the
curation job in webref to directly call `crawlSpecs`. Let's use the major
version bump opportunity to fix this. I added a `crawlList` exported function in
case someone actually used that entry point.