Fix exported functions crawlSpecs and getInterfaceTreeInfo (#1018)
The `getTreeInfo` entry point should have been renamed into
`getInterfaceTreeInfo`.

The `crawlSpecs` function targeted the `crawlList` crawl function instead of
`crawlSpecs`. I'm not clear that was intended and it would be easier for the
curation job in webref to directly call `crawlSpecs`. The `crawlList` function is
now hidden behind `crawlSpecs`, depending on function parameters.

We should probably try to get rid of `crawlList` in a future revision of Reffy.
Within Reffy, the function is merely useful for testing purpose.