Couple of bug fixes in crawler (#1019)
* Make "--use-crawl" option friendly to Windows line endings

Always a pain to tell Git under Windows not to convert `\n` to `\r\n` in text
files. This made the expansion function fail to detect the end of the header
and produce an invalid IDL extract when the `--use-crawl` option was used.

* Drop check on "spec.error" before crawl

Not sure why the check was there but nothing in the code sets the `error`
property before attempting to crawl the spec and when the `--use-crawl`
option is used, specs with crawling errors still need to be expanded as
error may only target one specific crawl module, and we should still try
to run post-processing modules.