Fix line ending detection in IDL extracts
The check on Windows line endings always returned true because indexOf returns
`-1` when no string is found. Now using `includes`.

This fix also strengthens the check to detect a double newline instead of
simple one for the unlikely case where extracts would use a mix of `\n` and
`\r\n`.