Extract CSS at-rules, move descriptors under at-rules (#1033)
This creates an `atrules` property in CSS extracts that contains at-rules
syntax as suggested in #1028. Descriptors associated with an at-rule (and all
selectors are associated with an at-rule in practice) now appear under the
at-rule definition.

The new `atrules` property contains a `value` property when a formal syntax
could be extracted from the spec, and a `descriptors` array that lists
descriptors associated with the at-rule (if any).

Breaking change: The former `descriptors` property that appeared at the root
level of the extract no longer exists. The same info can be found under the
`atrules` property.

This update also adjusts a couple of tests and adds two new tests on at-rules
extraction.

The code can only extract the formal syntax of an at-rule if it is defined in
a `<pre class="prod">` block.

As opposed to the structure proposed in #1028, the list of descriptors is not
indexed by the descriptor's name.

For instance, the `@container` at-rule would lead to:

```json
{
  "@container": {
    "value": "@container [ <container-name> ]? <container-condition> { <stylesheet> }",
    "descriptors": [
      {
        "name": "width",
        "for": "@container",
        "value": "<length>",
        "type": "range"
      },
      {
        "name": "height",
        "for": "@container",
        "value": "<length>",
        "type": "range"
      },
      {
        "name": "inline-size",
        "for": "@container",
        "value": "<length>",
        "type": "range"
      },
      {
        "name": "block-size",
        "for": "@container",
        "value": "<length>",
        "type": "range"
      },
      {
        "name": "aspect-ratio",
        "for": "@container",
        "value": "<ratio>",
        "type": "range"
      },
      {
        "name": "orientation",
        "for": "@container",
        "value": "portrait | landscape",
        "type": "discrete"
      }
    ]
  }
}
```