Events extraction: Reflect new URL of Service Workers spec (#1044)
ReSpec switched to the new URL of Service Workers (without level). This change
adds the new URLs as they are typically now used in specs such as Payment
Handler and the Push API to reference the "fire a functional event" concept.

The old "service-workers-1" URLs are probably still used in some specs, so let's
keep them for now.