Adjust check-missing-dfns script to handle at-rules
This makes the script look at at-rules and descriptors, following the new
structure returned by the crawler in the CSS extracts.

The script only reports a missing definition for an at-rule when the rule has
a formal syntax definition in the CSS extract (there are a number of specs that
extend an at-rule defined elsewhere and script would return them as missing
definitions without that check).

Fixes #1043.