Events post-processing: add bubblingPath property
This adds a `bubblingPath` property to the description of target interfaces
that contains, for bubbling events, the list of interfaces (at the root of the
inheritance chain) on which the event may bubble.

For instance, with that change, the definition of `fullscreenchange` becomes:

```json
{
  "type": "fullscreenchange",
  "interface": "Event",
  "targets": [
    {
      "target": "Element",
      "bubbles": true,
      "bubblingPath": [
        "Node",
        "Document",
        "Window"
      ]
    }
  ],
  "src": {
    "format": "dfn",
    "href": "https://fullscreen.spec.whatwg.org/#eventdef-document-fullscreenchange"
  },
  "href": "https://fullscreen.spec.whatwg.org/#eventdef-document-fullscreenchange"
}
```

Actual interfaces on which the event will bubble may well be interfaces that
inherit from the interfaces in `bubblingPath`. For instance, `fullscreenchange`
will actually bubble on `Element` (and not `Node`), `Document`... and not sure
the event actually fires on `ShadowRoot` or `Window` as the spec says that the
event handlers are not supported by these interfaces. Anyway, the point is: in
the DOM, we don't really know the exact list and can only report the core set of
interfaces that are on the bubbling path.

The order of the interfaces in the array is the order of the bubbling path.

This update avoids having to add a separate `bubbling-trees.json` file to the
events package, which would be hard to associate with the target interfaces in
any case (because one would need to look at the inheritance chain to understand
where the target interface fits in the bubbling tree). This is still imperfect
because it would be more interesting to have more precise interfaces in there,
but that's meant to be a step in the right direction.