Restructure links extracts to keep original links
Currently, reffy "canonicalize" urls before recording them in the links extracts.
While that canonicalization is useful to regroup links that points to the same spec, it loses critical information (e.g. for links to multipage specs) and produces information that does not really reflect the reality of the spec, making it hard to produce reliable analysis.

The patch instead records the original link untouched, and annotates it with the spec that can be identified after canonicalization.

This means the structure of links extract will change.