[references] Handle case when href attribute starts with whitespaces
The selector rule only selected `href` attribute that started with `http`.
There are a couple of occurrences in specs where the URL is prefixed with
whitespaces, which is technically correct.

This update rather looks for `://` somewhere in the attribute.

The update also completely removes the `url` property when no URL can be found
(instead of reporting an empty string).