[headings] Only report ID when it exists (and fix URL in that case)
Once in a while, especially in multi-page specs, a definition may be defined in
near the top of the page following a heading that does not have any ID. The
reported ID was an empty string in such cases, and the URL ended with `#null`.

This update drops the ID when it does not exist and the null fragment.