Fix spec post-processing when Reffy does not start with a crawl result (#1067)
The crawler assumed that it had access to the object describing post-processing
modules, whereas it only receives post-processing module names. This prevented
it from copying the results of some post-processing modules (`idlparsed` in
practice). Everything still worked when post-processing was run from a crawl
result, so this bug does not affect webref's use of Reffy in practice.