Add JSON schemas for extracts export validation function
This creates JSON schemas for all extracts and files generated by Reffy. A new
`getSchemaValidationFunction` function is now exported by Reffy to validate data
against the schemas. It is typically intended for use in Webref to validate
curated data before publication.

Some minor adjustments made along the way to make data more consistent and keep
schemas relatively simple:
- The dfns extractor now always returns a `heading` property, defaulting to the
top of the page
- The events extractor now gets rid of null properties altogether
- The events post-processor no longer outputs null href.

Existing tests on extractors were updated to also check the schema of the
extracted data. More tests could be added to check post-processors, and the
crawler as a whole.

Note this update also includes a minor bug fix in the sorting code of the events
post-processor so that it may run on not-yet-curated events extracts.