Re-write CSS extraction logic. Add values.
The CSS extraction logic was grounded on production rules, did not include
possible values of a property, did not know how to deal with functions and
types that are namespaced to a property or to another type, and did not
include selectors.

This update changes all of that:
- Selectors are now reported under a `selectors` property at the root level.
- Possible values that some definition may take are now reported under a
`values` property in the definition.
- Functions and types that are namespaced to some other definition are
included in the list of `values` of that definition.

Some breaking changes along the way:
1. Arrays are now used throughout instead of indexed objects, because that's
easier to deal with in code. It is easy to create an index from an array in
JavaScript if someone needs it.
2. Function names are no longer enclosed in `<` and `>` because they are not
defined in specs with these characters (as opposed to types). Beware though,
references to functions in value syntax do use enclosing `<` and `>`
characters.
3. The property `valuespaces` at the root level was replaced with a `values`
property (and an array is also used there). This `values` property lists both
`function` and `type` definitions that are not namespaced to anything in
particular. There are no `value` definitions at the root level because `value`
definitions are always namespaced to some other definition. And yes, naming is
hard.

Some additional non-breaking changes:
- To distinguish between `function`, `type` and `value` definitions listed in a
`values` property, definitions that appear in a `values` property have a `type`
property.
- Only namespaced values associated with a definition are listed under its
`values` property. Non-namespaced values are not. For instance, `<quote>` is
not listed as a value of the `<content-list>` type, even though its value
syntax references it. This avoids duplicating constructs in the extracts.
- Values are only listed under the deepest definition to which they apply. For
instance, `open-quote` is only listed as a value of `<quote>` but neither as a
value of the `<content-list>` type that references `<quote>` nor as a value of
the `content` property that references `<content-list>`. This is also to avoid
duplicating constructs in the extracts.