Handle case when `data-dfn-for` attribute is empty
An empty attribute made the code crash.

This happens for the second definition of `<size>` in:
https://drafts.csswg.org/css-images-3/#typedef-size%E2%91%A0

(FWIW, that definition is actually invalid because it re-defines a type that is
already defined in a production rule at the beginning of the section. This makes
the extraction throw an error for now as it can't yet deal with duplicates)