Ignore selectors definitions flagged as not exported
HTML and Fullscreen re-define selectors that are already defined in CSS
Selectors. There re-definition are internal to the spec. This update ignores
them when preparing the CSS extract.

Talking about exported definitions, all CSS properties, at-rules and descriptors
are currently defined with a `data-export` attribute. Most selectors and CSS
values (`function`, `type`, and `value`) are exported too, but 16 of them are
not. Private values are needed to make sense of other constructs, so these ones
do appear in the extracts.