Adjust CSS extract schemas and logic (#1125)
* Fix CSS extract schema: Values may not have "value"

Only CSS values of type `value` are guaranteed to have a `value` property,
values that are actually `type` or `function` may not.

* Fix CSS extract schema: An at-rule may have values

Most at-rules have only descriptors, but some do have direct values. That is
typically the case for the `@media` at-rule (with values such as `not` or
`all`). Adjusting the schema accodingly.

* CSS extraction: Allow selectors to have child selectors

Some selectors may have child selectors, as `::prefix` for `::first-letter` in
css-pseudo:
https://drafts.csswg.org/css-pseudo-4/#first-letter-pseudo

The extraction code wasn't expecting that.