CSS extracts: handle dfns with multiple linking texts
Via https://github.com/w3c/webref/issues/808

CSS 2 defines the `<identifier>` type with both a prose linking text
`identifiers` and the syntax linking text `<identifier>`. The code was not
expecting that and extracted `identifiers|<identifier>` as a result.

The code now extracts the linking text that looks like syntax if there is more
than one linking text defined. If none is found, the extraction currently throws
an error (to mean "case not supported yet"). We may want to be smarter about
this in the future... if that ever happens.