CSS extracts: fix function that looks for deepest construct
Via #1133.

There was a typo in the call to `checkChild` as the function does not take two
parameters. This made the code miss a number of values as it could not find
any "deepest construct" in the list of constructs that the value was for.

Also, the logic that detected cycles was totally wrong since it was filling
the `seen` array *after* the recursion. To keep things simple, the code simply
stops the recursion after 10 levels of recursion, which should be more than
enough.