CSS extract: Ignore informative code in `<dd>` element (#1139)
The extraction code assumed that code that appears in a `<dd>` element that
follows a `<dt>` definition describes the production value of the definition.
That seems rather bold. This update restricts it to normative code to fix #1138.

The whole assumption should probably be thoroughly re-evaluated though. The code
contains a comment to that effect in the meantime.