CSS extraction: only extract code with a "prod" class
The code considered that any kind of code appearing in a `<dd>` that followed
a definition was a production rule. This is incorrect as there are cases where
the prose uses code that is not meant to describe the value of the construct
defined.

This updates the logic to only consider code that has a `prod` class.

Fixes #1140