Extract element from MathML Core definitions (#1152)
* Extract element from MathML Core definitions

* Adjust crawl test not to expect an empty "elements"

New version of the "elements" extractor reports null when no elements could be
extracted, instead of an empty array. We're not very consistent in the way we
report lack of results for now, so both are fine, but crawl tests needed to be
adjusted accordingly.

* Fix check on SVG elements

That does not happen in the real SVG spec, but there could no elements defined
in element summary tables and some elements in definition tables. That typically
happens in tests.

* Adjust elements tests to pass the spec as argument

The extractor now makes use of the spec and expects it to be defined. The tests
needed to be adjusted accordingly.

A new test was also added on the extraction of MathML elements.