Associate element with interface in simple case
The elements extraction code did not have common logic to associate an element
`Xx` with its interface. If the code finds an interface whose name is
`HTMLXxElement`, it now automatically associates the element with it.

Typically useful for the `<portal>` element defined in "Portals" and the
`<model>` element defined in "The `<model>` element".