Filter out dfns defined in an "exclude" IDL block
Trusted Types has this weird IDL block where it defines DOMParser but does not
seem to mean it:
https://w3c.github.io/trusted-types/dist/spec/#integration-with-dom-parsing

The IDL block has an `exclude` class and is thus ignored by the IDL extraction
logic. This update completes the dfns extraction logic to also ignore terms
defined in such an IDL block. This avoids ending up with duplicated terms.