Add crude dfns patching mechanism (#1159)
* Filter out dfns defined in an "exclude" IDL block

Trusted Types has this weird IDL block where it defines DOMParser but does not
seem to mean it:
https://w3c.github.io/trusted-types/dist/spec/#integration-with-dom-parsing

The IDL block has an `exclude` class and is thus ignored by the IDL extraction
logic. This update completes the dfns extraction logic to also ignore terms
defined in such an IDL block. This avoids ending up with duplicated terms.

* Add post-processing module to patch dfns

This creates a crude patching mechanism to patch dfns directly after crawling a
spec. This can be used to drop some well-known duplicates that cause referencing
issues.

The module needs to be maintained over time, so goal is to only hardcode rules
when really needed, e.g. because the re-definition of a term causes actual
referencing issues.

The module runs at the spec level for simplicity. A more complete version could
run at the crawl level, detect duplicates and try to act on them. That would
require investing more energy into it :)