Skip extracts from old specs (#1174)
There are already cases (css-cascade notably) where the list of specs
contains an older level. This works fine because the older level does not
define something that further levels don't define (e.g. IDL). This update
tightens the code to make sure that older levels do not get considered when
looking for the latest version that passes a certain condition.

This will be particularly useful if we start listing previous levels of CSS
specs.