Release minor version of Reffy - v11.3.0
Dependencies bumped:
- Bump web-specs from 2.39.0 to 2.41.0 (#1177)
- Bump puppeteer from 19.4.1 to 19.5.0 (#1176)
- Bump commander from 9.4.1 to 9.5.0 (#1175)
- Bump respec from 32.6.0 to 32.6.1 (#1169)
- Bump rollup from 3.7.5 to 3.9.1 (#1170)
- Bump ajv from 8.11.2 to 8.12.0 (#1171)

Patches:
- Skip extracts from old specs (#1174)
- Fix data-lt extraction for CSS values (#1172)

The list of specs in web-specs reports a new `standing` property to flag specs
that have been discontinued, or that have not yet been fully adopted by a group.
The list now contains a few specs whose standing is either `pending` or
`discontinued`. Similarly, the list now contains a few older versions of CSS
specs.

These changes should not affect crawl results, but consumers may need to
filter out specs that are not in good standing depending on what they may want
to achieve with the data. Hence the minor version bump.