Extract headings without data-dfn-type that look like dfns (#1183)
We already consider that a `<dfn>` without a `data-dfn-type` has an implicit
`"data-dfn-type"="dfn"` attribute. This update extends that principle to
headings too, provided these headings have some other definition-related
attribute to make it clear that they want to be considered as definitions as
well.

This is typically needed for the definition of "transferable object" in HTML,
that has both a `data-export` and `data-lt` attribute but no `data-dfn-type`
attribute:
https://html.spec.whatwg.org/multipage/structured-data.html#transferable-objects

Without this update, Reffy would extract a private "transferable objects" term.
With this update, it extracts a public "transferable object" term as intended.

In practice, the only other heading across all specs that has a
definition-related attribute but no `data-dfn-type` is in the Reporting API:
https://w3c.github.io/reporting/#generate-report
(That definition is a bit weird because the spec uses `data-algorithm` where
we probably would prefer to see `data-lt`, but I propose to live with it for
now).