Accept singular for "new value" in partial CSS dfns
When a spec extends the values of a CSS property, the new values appear in a
"New Values" row in the definition table. The MathML Core specification
decided to use the singular form "New Value" for `text-transform` because it
only creates one new value. That seems fine and worth supporting in Reffy.
This update makes Reffy generate a proper `newValues` field even when singular
is used.

A couple of tests were added along the way, one for the new behavior, and one
for newValues in general as there was no actual test on that.